backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
openai==0.28.0
faiss-cpu==1.7.4
numpy==1.26.2
PyPDF2==3.0.1
pydantic==1.10.12
python-multipart==0.0.6
requests==2.31.0
pytest==7.4.2
sqlalchemy==2.0.41
alembic==1.11.1
